third_party/blink/web_tests/external/wpt/css/css-sizing/responsive-iframe/responsive-iframe-fallback.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] The fallback size is used
  assert_equals: expected 100 but got 150
Harness: the test ran to completion.
